initial commit of rewrite
